fix: improve scroll behavior
